requirements.txt
streamlit
pillow
pytesseract
opencv-python
SpeechRecognition
gTTS
